Extract OTA backend functionality into separate ota_base component
