+++ esphome/components/esphome/ota/ota_esphome.h
@@ -4,7 +4,7 @@
 #ifdef USE_OTA
 #include "esphome/core/helpers.h"
 #include "esphome/core/preferences.h"
-#include "esphome/components/ota/ota_backend.h"
+#include "esphome/components/ota/ota.h"
 #include "esphome/components/socket/socket.h"
 
 namespace esphome {
