+++ esphome/components/ota/ota.cpp
@@ -1,4 +1,4 @@
-#include "ota_backend.h"
+#include "ota.h"
 
 namespace esphome {
 namespace ota {
