+++ esphome/components/ota_base/ota_backend.cpp
@@ -1,0 +1,9 @@
+#include "ota_backend.h"
+
+namespace esphome {
+namespace ota_base {
+
+// The make_ota_backend() implementation is provided by each platform-specific backend
+
+}  // namespace ota_base
+}  // namespace esphome
