+++ esphome/components/ota_base/ota_backend_arduino_esp32.h
@@ -1,30 +1,24 @@
 #pragma once
-#ifdef USE_ARDUINO
-#ifdef USE_ESP8266
+#ifdef USE_ESP32_FRAMEWORK_ARDUINO
 #include "ota_backend.h"
 
 #include "esphome/core/defines.h"
-#include "esphome/core/macros.h"
+#include "esphome/core/helpers.h"
 
 namespace esphome {
-namespace ota {
+namespace ota_base {
 
-class ArduinoESP8266OTABackend : public OTABackend {
+class ArduinoESP32OTABackend : public OTABackend {
  public:
   OTAResponseTypes begin(size_t image_size) override;
   void set_update_md5(const char *md5) override;
   OTAResponseTypes write(uint8_t *data, size_t len) override;
   OTAResponseTypes end() override;
   void abort() override;
-#if USE_ARDUINO_VERSION_CODE >= VERSION_CODE(2, 7, 0)
-  bool supports_compression() override { return true; }
-#else
   bool supports_compression() override { return false; }
-#endif
 };
 
-}  // namespace ota
+}  // namespace ota_base
 }  // namespace esphome
 
-#endif
-#endif
+#endif  // USE_ESP32_FRAMEWORK_ARDUINO
