+++ esphome/components/ota_base/ota_backend_arduino_libretiny.cpp
@@ -1,20 +1,20 @@
-#ifdef USE_ESP32_FRAMEWORK_ARDUINO
+#ifdef USE_LIBRETINY
+#include "ota_backend_arduino_libretiny.h"
+#include "ota_backend.h"
+
 #include "esphome/core/defines.h"
 #include "esphome/core/log.h"
-
-#include "ota_backend.h"
-#include "ota_backend_arduino_esp32.h"
 
 #include <Update.h>
 
 namespace esphome {
-namespace ota {
+namespace ota_base {
 
-static const char *const TAG = "ota.arduino_esp32";
+static const char *const TAG = "ota.arduino_libretiny";
 
-std::unique_ptr<ota::OTABackend> make_ota_backend() { return make_unique<ota::ArduinoESP32OTABackend>(); }
+std::unique_ptr<OTABackend> make_ota_backend() { return make_unique<ArduinoLibreTinyOTABackend>(); }
 
-OTAResponseTypes ArduinoESP32OTABackend::begin(size_t image_size) {
+OTAResponseTypes ArduinoLibreTinyOTABackend::begin(size_t image_size) {
   bool ret = Update.begin(image_size, U_FLASH);
   if (ret) {
     return OTA_RESPONSE_OK;
@@ -29,9 +29,9 @@ OTAResponseTypes ArduinoESP32OTABackend::begin(size_t image_size) {
   return OTA_RESPONSE_ERROR_UNKNOWN;
 }
 
-void ArduinoESP32OTABackend::set_update_md5(const char *md5) { Update.setMD5(md5); }
+void ArduinoLibreTinyOTABackend::set_update_md5(const char *md5) { Update.setMD5(md5); }
 
-OTAResponseTypes ArduinoESP32OTABackend::write(uint8_t *data, size_t len) {
+OTAResponseTypes ArduinoLibreTinyOTABackend::write(uint8_t *data, size_t len) {
   size_t written = Update.write(data, len);
   if (written == len) {
     return OTA_RESPONSE_OK;
@@ -43,7 +43,7 @@ OTAResponseTypes ArduinoESP32OTABackend::write(uint8_t *data, size_t len) {
   return OTA_RESPONSE_ERROR_WRITING_FLASH;
 }
 
-OTAResponseTypes ArduinoESP32OTABackend::end() {
+OTAResponseTypes ArduinoLibreTinyOTABackend::end() {
   if (Update.end()) {
     return OTA_RESPONSE_OK;
   }
@@ -54,9 +54,9 @@ OTAResponseTypes ArduinoESP32OTABackend::end() {
   return OTA_RESPONSE_ERROR_UPDATE_END;
 }
 
-void ArduinoESP32OTABackend::abort() { Update.abort(); }
+void ArduinoLibreTinyOTABackend::abort() { Update.abort(); }
 
-}  // namespace ota
+}  // namespace ota_base
 }  // namespace esphome
 
-#endif  // USE_ESP32_FRAMEWORK_ARDUINO
+#endif  // USE_LIBRETINY
