+++ esphome/components/ota_base/ota_backend_arduino_libretiny.h
@@ -1,15 +1,13 @@
 #pragma once
-#ifdef USE_ARDUINO
-#ifdef USE_RP2040
+#ifdef USE_LIBRETINY
 #include "ota_backend.h"
 
 #include "esphome/core/defines.h"
-#include "esphome/core/macros.h"
 
 namespace esphome {
-namespace ota {
+namespace ota_base {
 
-class ArduinoRP2040OTABackend : public OTABackend {
+class ArduinoLibreTinyOTABackend : public OTABackend {
  public:
   OTAResponseTypes begin(size_t image_size) override;
   void set_update_md5(const char *md5) override;
@@ -19,8 +17,7 @@ class ArduinoRP2040OTABackend : public OTABackend {
   bool supports_compression() override { return false; }
 };
 
-}  // namespace ota
+}  // namespace ota_base
 }  // namespace esphome
 
-#endif  // USE_RP2040
-#endif  // USE_ARDUINO
+#endif  // USE_LIBRETINY
