+++ esphome/components/ota_base/ota_backend_arduino_rp2040.cpp
@@ -10,11 +10,11 @@
 #include <Updater.h>
 
 namespace esphome {
-namespace ota {
+namespace ota_base {
 
 static const char *const TAG = "ota.arduino_rp2040";
 
-std::unique_ptr<ota::OTABackend> make_ota_backend() { return make_unique<ota::ArduinoRP2040OTABackend>(); }
+std::unique_ptr<OTABackend> make_ota_backend() { return make_unique<ArduinoRP2040OTABackend>(); }
 
 OTAResponseTypes ArduinoRP2040OTABackend::begin(size_t image_size) {
   bool ret = Update.begin(image_size, U_FLASH);
@@ -68,7 +68,7 @@ void ArduinoRP2040OTABackend::abort() {
   rp2040::preferences_prevent_write(false);
 }
 
-}  // namespace ota
+}  // namespace ota_base
 }  // namespace esphome
 
 #endif  // USE_RP2040
