+++ esphome/components/ota_base/ota_backend_esp_idf.cpp
@@ -12,9 +12,9 @@
 #endif
 
 namespace esphome {
-namespace ota {
+namespace ota_base {
 
-std::unique_ptr<ota::OTABackend> make_ota_backend() { return make_unique<ota::IDFOTABackend>(); }
+std::unique_ptr<OTABackend> make_ota_backend() { return make_unique<IDFOTABackend>(); }
 
 OTAResponseTypes IDFOTABackend::begin(size_t image_size) {
   this->partition_ = esp_ota_get_next_update_partition(nullptr);
@@ -111,6 +111,6 @@ void IDFOTABackend::abort() {
   this->update_handle_ = 0;
 }
 
-}  // namespace ota
+}  // namespace ota_base
 }  // namespace esphome
 #endif
