+++ esphome/components/speaker/media_player/speaker_media_player.cpp
@@ -6,7 +6,7 @@
 
 #include "esphome/components/audio/audio.h"
 #ifdef USE_OTA
-#include "esphome/components/ota/ota_backend.h"
+#include "esphome/components/ota/ota.h"
 #endif
 
 namespace esphome {
